Sim updates, input hooks, brush size
Removed random sampled cell updates, as this was a source of issue.
Added a garbage input hook system. They serve as a way to call a function elsewhere immediately after the event is fired.
Added a brush size system, using input hooks. for now the brush is square, has no indicator and is output into terminal.